LanguageController made more generic.
The LanguageController trait is now more generic, and now uses a single route file to
make the language change.

/language/english
/language/welsh

The language string is processed by the switchToLanguage public method which checks the
languageMap object for a mapping of the language to a Lang object.

By default, the library provides English and Welsh support, but new languages can be added
by creating a new controller object that extends LanguageController and then defining
a custom languageMap override.

    override def languageMap = Map("spanish" -> Lang("es"))